requirements.txt
# Core ML / DL libs
tensorflow==2.19.0
keras==3.9.2
numpy==1.24.4
h5py==3.13.0
pillow==11.2.1
opencv-python==4.7.0.72
scikit-learn==1.2.2
matplotlib==3.7.1

# Django and REST Framework
Django==5.2.1
djangorestframework==3.16.0

# Other useful libs
absl-py==2.2.2
asgiref==3.8.1
gast==0.6.0
grpcio==1.71.0
protobuf==5.29.4
werkzeug==3.1.3
termcolor==3.0.1
typing_extensions==4.13.2
tzdata==2025.2

# Networking / requests
requests==2.32.3
urllib3==2.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
